Long

+++ app/src/main/java/com/example/canvastaks/data/model/TaskModel.kt
@@ -6,7 +6,8 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "task")
 data class TaskModel (
     @PrimaryKey(autoGenerate = true)
-    val id: Int = 0,
+    val id: Long = 0,
+
     val tasks: MutableList<TaskNodeModel>,
     val taskTitle: String
 )
